+++ pyanalyze/checker.py
@@ -12,8 +12,8 @@ from .value import TypedValue
 from .arg_spec import ArgSpecCache
 from .config import Config
 from .reexport import ImplicitReexportTracker
-from .safe import is_typing_name
-from .type_object import TypeObject
+from .safe import is_instance_of_typing_name, is_typing_name, safe_getattr
+from .type_object import TypeObject, get_mro
 
 
 @dataclass
@@ -62,8 +62,8 @@ class Checker:
             return TypeObject(typ, self.get_additional_bases(typ))
         else:
             additional_bases = self.get_additional_bases(typ)
-            is_protocol = self.arg_spec_cache.ts_finder.is_protocol(typ)
-            if is_protocol:
+            # Is it marked as a protocol in stubs? If so, use the stub definition.
+            if self.arg_spec_cache.ts_finder.is_protocol(typ):
                 bases = self._get_typeshed_bases(typ)
                 return TypeObject(
                     typ,
@@ -71,6 +71,20 @@ class Checker:
                     is_protocol=True,
                     protocol_members=self._get_protocol_members(bases),
                 )
+            # Is it a protocol at runtime?
+            if is_instance_of_typing_name(typ, "_ProtocolMeta") and safe_getattr(
+                typ, "_is_protocol", False
+            ):
+                bases = get_mro(typ)
+                members = set(
+                    itertools.chain.from_iterable(
+                        _extract_protocol_members(base) for base in bases
+                    )
+                )
+                return TypeObject(
+                    typ, additional_bases, is_protocol=True, protocol_members=members
+                )
+
             return TypeObject(typ, additional_bases)
 
     def _get_typeshed_bases(self, typ: Union[type, str]) -> Set[Union[type, str]]:
@@ -99,3 +113,45 @@ class Checker:
         finally:
             new_pair = self.assumed_compatibilities.pop()
             assert pair == new_pair
+
+
+EXCLUDED_PROTOCOL_MEMBERS = {
+    "__abstractmethods__",
+    "__annotations__",
+    "__dict__",
+    "__doc__",
+    "__init__",
+    "__new__",
+    "__module__",
+    "__parameters__",
+    "__subclasshook__",
+    "__weakref__",
+    "_abc_impl",
+    "_abc_cache",
+    "_is_protocol",
+    "__next_in_mro__",
+    "_abc_generic_negative_cache_version",
+    "__orig_bases__",
+    "__args__",
+    "_abc_registry",
+    "__extra__",
+    "_abc_generic_negative_cache",
+    "__origin__",
+    "__tree_hash__",
+    "_gorg",
+}
+
+
+def _extract_protocol_members(typ: type) -> Set[str]:
+    if (
+        typ is object
+        or is_typing_name(typ, "Generic")
+        or is_typing_name(typ, "Protocol")
+    ):
+        return set()
+    members = set()
+    if hasattr(typ, "__dict__"):
+        members |= set(typ.__dict__) - EXCLUDED_PROTOCOL_MEMBERS
+    if hasattr(typ, "__annotations__"):
+        members |= set(typ.__annotations__)
+    return members