+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add support for runtime Protocols (#311)
 - Stop inferring `Any` for non-runtime checkable Protocols on Python 3.6 and 3.7 (#310)
 - Fix false positive where `multiprocessing.Pool.map_async`
   was identified as an asynq method (#306)